refactor: 폴더 및 파일 이름 변경
